// src/types/api/bindings/ApiCreateCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { ApiCreateCommandOption } from "./ApiCreateCommandOption";

/**
 * Publicly accessible representation of a Discord command
 */
export type ApiCreateCommand = { type: number, name: string | null, name_localizations: { [key in string]?: string }, description: string | null, description_localizations: { [key in string]?: string }, integration_types: number, nsfw: boolean, options: Array<ApiCreateCommandOption>, };